Upgrade Jmix to 3.0 #167 (#284)
1. Update Jmix version to 3.0.1 (Spring Boot 4, Vaadin Flow 25)
2. Adapt CSS styles to Vaadin Flow 25
3. Update action classes: add generic type, e.g. `SecuredBaseAction<RetryJobAction>` instead of `SecuredBaseAction`
4. Replace @Nullable annotations and use Jspecify annotations
5. Update security configuration:
   - Migrate OidcSecurityConfiguration.java
   - Add public access to frontend resources like CSS styles
6. Get rid of using deprecated `org.apache.commons.lang3.StringUtils.equals`-like methods
7. Update UI tests and adapt for changes related to Vaadin Flow 25
8. Migrate from Jackson 2 to Jackson 3. Leave Jackson 2 for compatibility with a new Camunda client.
9. Migrate to new Camunda 7 client, update CamundaFeignConfiguration
10. Replace standard Jmix pagination query parameter binder with custom ControlAbstractUrlQueryParametersBinder to fix clearing filter values on list views with delayed load

+++ src/main/java/io/flowset/control/entity/ExternalTaskData.java
@@ -45,7 +45,7 @@ public class ExternalTaskData {
 
     protected String topicName;
 
-    protected Integer priority;
+    protected Long priority;
 
     protected String businessKey;
 
@@ -65,11 +65,11 @@ public class ExternalTaskData {
         this.businessKey = businessKey;
     }
 
-    public Integer getPriority() {
+    public Long getPriority() {
         return priority;
     }
 
-    public void setPriority(Integer priority) {
+    public void setPriority(Long priority) {
         this.priority = priority;
     }
